LHS pattern annotations can contain chunks
